[feat]: [CI-3005]: Support for plugin and runtest with AWS infra (#29112)
* [CI-3005]: Support for plugin and runtest runner integration

* [feat]:[CI-3005]: Support for plugin and runtest with AWS infra

* fix Ut

* fix checkstyle

* fix checkstyle

* fix hash

* fix checkstyle

* fix checkstyle

+++ 950-delegate-tasks-beans/src/main/java/io/harness/delegate/beans/ci/vm/steps/VmPluginStep.java
@@ -17,7 +17,7 @@ public class VmPluginStep implements VmStepInfo {
 
   private Map<String, String> envVariables;
   private VmUnitTestReport unitTestReport;
-  private int timeoutSecs;
+  private long timeoutSecs;
 
   @Override
   public VmStepInfo.Type getType() {
