+++ 950-delegate-tasks-beans/src/main/java/io/harness/delegate/beans/ci/vm/steps/VmRunTestStep.java
@@ -28,7 +28,7 @@ public class VmRunTestStep implements VmStepInfo {
   private Map<String, String> envVariables;
   private List<String> outputVariables;
   private VmUnitTestReport unitTestReport;
-  private int timeoutSecs;
+  private long timeoutSecs;
 
   @Override
   public VmStepInfo.Type getType() {
